Split SocketServer.
