+++ tests/StatsdClient.Tests/utils/SocketServer.cs
@@ -1,22 +1,13 @@
-using System;
-using System.Collections.Generic;
 using System.Net;
 using System.Net.Sockets;
-using System.Threading;
-using System.Threading.Tasks;
 using Mono.Unix;
 using StatsdClient;
 
 namespace Tests.Utils
 {
-    internal class SocketServer : IDisposable
+    internal class SocketServer : AbstractServer
     {
         private readonly Socket _server;
-        private readonly Task _receiver;
-        private readonly ManualResetEventSlim _serverStop = new ManualResetEventSlim(false);
-        private readonly List<string> _messagesReceived = new List<string>();
-
-        private volatile bool _shutdown = false;
 
         public SocketServer(StatsdConfig config)
         {
@@ -40,46 +31,24 @@ namespace Tests.Utils
 
             _server.ReceiveTimeout = 1000;
             _server.Bind(endPoint);
-            _receiver = Task.Run(() => ReadFromServer(bufferSize));
+            Start(bufferSize);
         }
 
-        public void Dispose()
+        public override void Dispose()
         {
-            Stop();
+            base.Dispose();
             _server.Dispose();
         }
 
-        public List<string> Stop()
+        protected override int? Read(byte[] buffer)
         {
-            if (!_shutdown)
+            try
             {
-                _shutdown = true;
-                _serverStop.Wait();
+                return _server.Receive(buffer);
             }
-
-            return _messagesReceived;
-        }
-
-        private void ReadFromServer(int bufferSize)
-        {
-            var buffer = new byte[bufferSize];
-
-            while (true)
+            catch (SocketException e) when (e.SocketErrorCode == SocketError.TimedOut)
             {
-                try
-                {
-                    var count = _server.Receive(buffer);
-                    var message = System.Text.Encoding.UTF8.GetString(buffer, 0, count);
-                    _messagesReceived.AddRange(message.Split("\n", StringSplitOptions.RemoveEmptyEntries));
-                }
-                catch (SocketException e) when (e.SocketErrorCode == SocketError.TimedOut)
-                {
-                    if (_shutdown)
-                    {
-                        _serverStop.Set();
-                        return;
-                    }
-                }
+                return null;
             }
         }
     }
